requirements.txt
beautifulsoup4>=4.0,<5.0
click>=8.0,<9.0
pandas>=2.0,<3.0
pandas-stubs>=2.0,<3.0
python-dotenv>=0.21.0,<0.22.0
requests>=2.0,<3.0
requests-cache>=0.9.0,<0.10.0
SQLAlchemy>=2.0,<3.0
tqdm>=4.0,<5.0
types-requests>=2.0,<3.0
typing-extensions>=4.0,<5.0
git+https://github.com/ranaroussi/yfinance.git@0.2.50
